Hacked jade/Styl
Resetting various configurations after changing username. Also, there
are some real sloppy jade and stylus files that need assistance. I'll
look for it to your demo and commits or comments. I know it's a mess –
but steep learning curve for three revised languages and all sorts of
functions and features a completely new to me. Such as the markdown.